Remove Ractor data from 20251001-20251123
During this period the ruby-bench suite was running the ractor
benchmarks without the --no-pinning option. This resulted in Ractor
benchmarks being pinned to a core, artificially constraining
performance.

The fix went out on Nov 21, but new data was delayed due to incorrectly
build docker containers. In the meantime older dates were picked up by
the backfill so dates have been backfilled prior to Ocober 1.

The docker image fixes went out on Nov 23r, resulting in correct data
being generated again from Nov 24 onwards.

Now that we've verified that the onward jobs and the backfilled jobs are
correct we can remove the misleading data to enable data for these days
to be regenerated by the backfill job.