Mark broken Rubies as broken
Docker images containing broken ruby executables have been shipped to
the container image registries. If this happens then add-ruby will fail
to add the sha to the rubies.yml file.

Because add-ruby attempts to add all missing rubies, a single failure
will prevent it from working again. This commit ensures that broken
rubies are "marked" in the rubies.yml file (as having a nil sha), and
future runs of add_ruby will skip attempting to build with a broken
container.